avoid printing control characters from goal (#1585)
Avoid printing control characters as part of the goal output
